[refactor] #200 이벤트 공통 서비스 메소드 이동
