+++ src/main/java/kyonggi/bookslyserver/domain/shop/service/ShopService.java
@@ -11,6 +11,7 @@ import kyonggi.bookslyserver.domain.shop.entity.Menu.Menu;
 import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
 import kyonggi.bookslyserver.domain.shop.entity.Shop.ShopImage;
 import kyonggi.bookslyserver.domain.shop.repository.BusinessScheduleRepository;
+import kyonggi.bookslyserver.domain.shop.repository.CategoryRepository;
 import kyonggi.bookslyserver.domain.shop.repository.ShopImageRepository;
 import kyonggi.bookslyserver.domain.shop.repository.ShopRepository;
 import kyonggi.bookslyserver.domain.user.entity.ShopOwner;
@@ -41,6 +42,11 @@ public class ShopService {
     private final ShopImageRepository shopImageRepository;
 
     private final ShopOwnerRepository shopOwnerRepository;
+
+    private final CategoryRepository categoryRepository;
+
+    private final String ALL_REGION = "전체";
+
 
 
     public ShopUserReadOneDto findOne(Long id){
@@ -163,6 +169,36 @@ public class ShopService {
             throw new InvalidValueException(NOT_OWNER_OF_SHOP);
         }
         return shop;
+    }
+
+    private List<Shop> filterShopsByAddress(List<Shop> shops, String region) {
+        String[] address = region.split(" ");
+
+        if (address[2].equals(ALL_REGION)) {
+            shops = shopRepository.findShopsByOneAndTwoAddress(address[0], address[1], shops);
+        }else {
+            shops = shopRepository.findShopsByAddress(address[0], address[1], address[2], shops);
+        }
+        return shops;
+    }
+
+    public List<Shop> getRegionFilteredShops(List<String> regions, List<Shop> shops) {
+        List<Shop> filteredShops = new ArrayList<>();
+        regions.stream().forEach(region ->
+                filteredShops.addAll(filterShopsByAddress(shops, region)));
+        return filteredShops;
+    }
+
+    private void validateCategoriesExist(List<Long> categoryIds) {
+        List<Long> existingCategoryIds = categoryRepository.findByCategoryIds(categoryIds);
+        if (existingCategoryIds.size() != categoryIds.size()) {
+            throw new InvalidValueException(ErrorCode.CATEGORY_NOT_FOUNT);
+        }
+    }
+
+    public List<Shop> getCategoryFilteredShops(List<Long> categories, List<Shop> shops) {
+        validateCategoriesExist(categories);
+        return shopRepository.findByCategoryIds(categories,shops);
     }
